Added checks

+++ structure/subscriberManagement.js
@@ -4,22 +4,43 @@ const logger = require('../util/logger.js');
 
 const subscriberManagement = {
 	addSubscriber: data => new Promise((resolve, reject) => {
-		db.addSubscriber(data).then(subscriber => {
-			logger.info('[subscriberManagement - addSubscriber(data): 1]\n', subscriber);
-			resolve(subscriber);
+		db.getSubscriberById(data.id).then(data => {
+			if (data) {
+				resolve();
+				return;
+			}
+			db.addSubscriber(data).then(subscriber => {
+				logger.info('[subscriberManagement - addSubscriber(data): 2]\n', subscriber);
+				resolve(subscriber);
+			}).catch(err => {
+				logger.error('[subscriberManagement - addSubscriber(data): 2]\n', err);
+				reject(err);
+			});
 		}).catch(err => {
 			logger.error('[subscriberManagement - addSubscriber(data): 1]\n', err);
 			reject(err);
 		});
 	}),
 	removeSubscriber: id => new Promise((resolve, reject) => {
-		db.removeSubscriber(id).then(subscriber => {
-			logger.info('[subscriberManagement - removeSubscriber(id)]\n', subscriber);
-			resolve(subscriber);
+		db.getSubscriberById(id).then(data => {
+			if (!data) {
+				resolve();
+				return;
+			}
+			db.removeSubscriber(id).then(subscriber => {
+				logger.info('[subscriberManagement - removeSubscriber(id): 2]\n', subscriber);
+				resolve(subscriber);
+			}).catch(err => {
+				logger.info('[subscriberManagement - removeSubscriber(id): 2]\n', err);
+				reject(err);
+			});
 		}).catch(err => {
-			logger.info('[subscriberManagement - removeSubscriber(id)]\n', err);
+			logger.info('[subscriberManagement - removeSubscriber(id): 1]\n', err);
 			reject(err);
 		});
+	}),
+	getSubscribers: () => new Promise((resolve, reject) => {
+
 	})
 };
